+++ src/theme/foundations/colors.ts
@@ -1,0 +1,30 @@
+export const colors = {
+	baseColors: {
+		grayScaleA: "#ffffff",
+		grayScaleB: "#e0e0e0",
+		grayScaleC: "#666666",
+		grayScaleD: "#f7f7f7",
+		grayScaleE: "#212121",
+	},
+
+	typeColors: {
+		normal: "#aaa67f",
+		ground: "#dec16b",
+		electric: "#f9cf30",
+		rock: "#b69e31",
+		dark: "#75574c",
+		fire: "#f57d31",
+		fairy: "#e69eac",
+		psychic: "#fb5584",
+		fighting: "#c12239",
+		poison: "#a43e9e",
+		dragon: "#7037ff",
+		ghost: "#70559b",
+		flying: "#a891ec",
+		steel: "#b7b9d0",
+		water: "#6493eb",
+		ice: "#9ad6df",
+		grass: "#74cb48",
+		bug: "#a7b723",
+	},
+};
