+++ src/theme/foundations/fonts.ts
@@ -1,0 +1,38 @@
+export const fonts = {
+	fontFamily: "Poppins, sans-serif",
+
+	fontWeight: {
+		regular: "400",
+		bold: "700",
+	},
+
+	heading: {
+		textSize: {
+			s: "1.8rem",
+			m: "2.4rem",
+			l: "3.2rem",
+		},
+
+		textHeight: {
+			s: "2.8rem",
+			m: "3.2rem",
+			l: "3.6rem",
+		},
+	},
+
+	body: {
+		textSize: {
+			xs: "1.2rem",
+			s: "1.4rem",
+			m: "1.6rem",
+			l: "1.8rem",
+		},
+
+		textHeight: {
+			xs: "1.6rem",
+			s: "2rem",
+			m: "2.4rem",
+			l: "2.4rem",
+		},
+	},
+};
