+++ src/theme/foundations/spaces.ts
@@ -1,0 +1,9 @@
+export const spaces = {
+	xxs: "0.4rem",
+	xs: "0.8rem",
+	s: "1.2rem",
+	m: "1.6rem",
+	l: "2.4rem",
+	xl: "3.2rem",
+	xxl: "4.8rem",
+};
