+++ tsconfig.json
@@ -20,6 +20,6 @@
       "@/*": ["src/*"]
     },
   },
-  "include": ["src", "stories", "tests/Tag.test.tsx", "tests/Header.test.tsx"],
+  "include": ["src", "stories", "tests" ],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
